fix: resolve review blockers from PR #23
1. Break circular dependency: move video_player_get_slow_level and
   video_player_calc_late_window_ms from layout module to playback module.
   Layout no longer depends on playback, and playback self-contains its
   timing-derived helpers.

2. Fix misplaced responsibility: move video_player_has_renderer from
   decode module to video_player.c core as a static function. It operates
   on render_mutex/renderer which are core concerns, not decode resources.

3. Fix rewind_needs_resync data race: add state_mutex lock around the
   read of player->rewind_needs_resync in video_player_should_drop_late_frame().
   The write side already holds state_mutex, so the read must too.

4. Consolidate duplicate typedef: centralize typedef RenderedFrame VideoFrame
   in video_player.h and remove 4 local copies in video_player.c,
   video_player_playback.c, video_player_seek.c, and test_video_player.c.

All 272 tests pass.
